refactor(aggregator): move builders function to `protocol` submodule
