+++ mithril-aggregator/src/dependency_injection/builder/mod.rs
@@ -3,7 +3,6 @@ mod protocol;
 mod support;
 
 use anyhow::Context;
-use semver::Version;
 use slog::{debug, Logger};
 use std::{collections::BTreeSet, path::PathBuf, sync::Arc};
 use tokio::{
@@ -21,30 +20,21 @@ use mithril_common::{
     cardano_transactions_preloader::{
         CardanoTransactionsPreloader, CardanoTransactionsPreloaderActivation,
     },
-    certificate_chain::{CertificateVerifier, MithrilCertificateVerifier},
+    certificate_chain::CertificateVerifier,
     chain_observer::{CardanoCliRunner, ChainObserver, ChainObserverBuilder, FakeObserver},
     chain_reader::{ChainBlockReader, PallasChainReader},
-    crypto_helper::{
-        MKTreeStoreInMemory, ProtocolGenesisSigner, ProtocolGenesisVerificationKey,
-        ProtocolGenesisVerifier,
-    },
+    crypto_helper::ProtocolGenesisVerifier,
     digesters::{
         cache::ImmutableFileDigestCacheProvider, CardanoImmutableDigester,
         DumbImmutableFileObserver, ImmutableDigester, ImmutableFileObserver,
         ImmutableFileSystemObserver,
     },
-    entities::{CompressionAlgorithm, Epoch, SignedEntityTypeDiscriminants},
+    entities::{Epoch, SignedEntityTypeDiscriminants},
     era::{
         adapters::{EraReaderAdapterBuilder, EraReaderDummyAdapter},
         EraChecker, EraMarker, EraReader, EraReaderAdapter, SupportedEra,
     },
-    signable_builder::{
-        CardanoDatabaseSignableBuilder, CardanoImmutableFilesFullSignableBuilder,
-        CardanoStakeDistributionSignableBuilder, CardanoTransactionsSignableBuilder,
-        MithrilSignableBuilderService, MithrilStakeDistributionSignableBuilder,
-        SignableBuilderService, SignableBuilderServiceDependencies, SignableSeedBuilder,
-        TransactionsImporter,
-    },
+    signable_builder::{SignableBuilderService, SignableSeedBuilder, TransactionsImporter},
     signed_entity_type_lock::SignedEntityTypeLock,
     MithrilTickerService, TickerService,
 };
@@ -55,44 +45,26 @@ use mithril_persistence::{
 
 use super::{DependenciesBuilderError, EpochServiceWrapper, Result};
 use crate::{
-    artifact_builder::{
-        AncillaryArtifactBuilder, AncillaryFileUploader, CardanoDatabaseArtifactBuilder,
-        CardanoImmutableFilesFullArtifactBuilder, CardanoStakeDistributionArtifactBuilder,
-        CardanoTransactionsArtifactBuilder, DigestArtifactBuilder, DigestFileUploader,
-        ImmutableArtifactBuilder, ImmutableFilesUploader, MithrilStakeDistributionArtifactBuilder,
-    },
     configuration::ExecutionEnvironment,
     database::repository::{
-        BufferedSingleSignatureRepository, CertificatePendingRepository, CertificateRepository,
-        EpochSettingsStore, ImmutableFileDigestRepository, OpenMessageRepository,
-        SignedEntityStore, SignedEntityStorer, SignerRegistrationStore, SignerStore,
-        SingleSignatureRepository, StakePoolStore,
+        CertificatePendingRepository, CertificateRepository, EpochSettingsStore,
+        ImmutableFileDigestRepository, OpenMessageRepository, SignedEntityStore,
+        SignedEntityStorer, SignerRegistrationStore, SignerStore, StakePoolStore,
     },
     entities::AggregatorEpochSettings,
     event_store::{EventMessage, EventStore, TransmitterService},
-    file_uploaders::{
-        CloudRemotePath, FileUploadRetryPolicy, FileUploader, GcpBackendUploader, GcpUploader,
-        LocalUploader,
-    },
-    http_server::{
-        routes::router::{self, RouterConfig, RouterState},
-        CARDANO_DATABASE_DOWNLOAD_PATH,
-    },
+    file_uploaders::FileUploader,
+    http_server::routes::router::{self, RouterConfig, RouterState},
     services::{
-        AggregatorSignableSeedBuilder, AggregatorUpkeepService, BufferedCertifierService,
-        CardanoTransactionsImporter, CertifierService, CompressedArchiveSnapshotter,
-        DumbSnapshotter, EpochServiceDependencies, MessageService, MithrilCertifierService,
-        MithrilEpochService, MithrilMessageService, MithrilProverService,
-        MithrilSignedEntityService, MithrilStakeDistributionService, ProverService,
-        SignedEntityService, SignedEntityServiceArtifactsDependencies, Snapshotter,
-        SnapshotterCompressionAlgorithm, StakeDistributionService, UpkeepService, UsageReporter,
+        AggregatorUpkeepService, CertifierService, EpochServiceDependencies, MessageService,
+        MithrilEpochService, MithrilMessageService, MithrilStakeDistributionService, ProverService,
+        SignedEntityService, Snapshotter, StakeDistributionService, UpkeepService, UsageReporter,
     },
     store::CertificatePendingStorer,
     tools::{CExplorerSignerRetriever, GenesisToolsDependency, SignersImporter},
     AggregatorConfig, AggregatorRunner, AggregatorRuntime, Configuration, DependencyContainer,
-    DumbUploader, EpochSettingsStorer, ImmutableFileDigestMapper, LocalSnapshotUploader,
-    MetricsService, MithrilSignerRegisterer, MultiSigner, MultiSignerImpl,
-    SingleSignatureAuthenticator, SnapshotUploaderType, VerificationKeyStorer,
+    EpochSettingsStorer, ImmutableFileDigestMapper, MetricsService, MithrilSignerRegisterer,
+    MultiSigner, SingleSignatureAuthenticator, VerificationKeyStorer,
 };
 
 const SQLITE_FILE: &str = "aggregator.sqlite3";
@@ -465,69 +437,6 @@ impl DependenciesBuilder {
         Ok(self.stake_store.as_ref().cloned().unwrap())
     }
 
-    async fn build_snapshot_uploader(&mut self) -> Result<Arc<dyn FileUploader>> {
-        let logger = self.root_logger();
-        if self.configuration.environment == ExecutionEnvironment::Production {
-            match self.configuration.snapshot_uploader_type {
-                SnapshotUploaderType::Gcp => {
-                    let allow_overwrite = true;
-                    let remote_folder_path = CloudRemotePath::new("cardano-immutable-files-full");
-
-                    Ok(Arc::new(
-                        self.build_gcp_uploader(remote_folder_path, allow_overwrite)?,
-                    ))
-                }
-                SnapshotUploaderType::Local => {
-                    let snapshot_artifacts_dir = self
-                        .configuration
-                        .get_snapshot_dir()?
-                        .join(SNAPSHOT_ARTIFACTS_DIR);
-                    std::fs::create_dir_all(&snapshot_artifacts_dir).map_err(|e| {
-                        DependenciesBuilderError::Initialization {
-                            message: format!(
-                                "Cannot create '{snapshot_artifacts_dir:?}' directory."
-                            ),
-                            error: Some(e.into()),
-                        }
-                    })?;
-
-                    Ok(Arc::new(LocalSnapshotUploader::new(
-                        self.configuration.get_server_url()?,
-                        &snapshot_artifacts_dir,
-                        logger,
-                    )))
-                }
-            }
-        } else {
-            Ok(Arc::new(DumbUploader::new(FileUploadRetryPolicy::never())))
-        }
-    }
-
-    /// Get a [FileUploader]
-    pub async fn get_snapshot_uploader(&mut self) -> Result<Arc<dyn FileUploader>> {
-        if self.snapshot_uploader.is_none() {
-            self.snapshot_uploader = Some(self.build_snapshot_uploader().await?);
-        }
-
-        Ok(self.snapshot_uploader.as_ref().cloned().unwrap())
-    }
-
-    async fn build_multi_signer(&mut self) -> Result<Arc<dyn MultiSigner>> {
-        let multi_signer =
-            MultiSignerImpl::new(self.get_epoch_service().await?, self.root_logger());
-
-        Ok(Arc::new(multi_signer))
-    }
-
-    /// Get a configured multi signer
-    pub async fn get_multi_signer(&mut self) -> Result<Arc<dyn MultiSigner>> {
-        if self.multi_signer.is_none() {
-            self.multi_signer = Some(self.build_multi_signer().await?);
-        }
-
-        Ok(self.multi_signer.as_ref().cloned().unwrap())
-    }
-
     async fn build_certificate_pending_storer(
         &mut self,
     ) -> Result<Arc<dyn CertificatePendingStorer>> {
@@ -862,113 +771,6 @@ impl DependenciesBuilder {
         Ok(self.immutable_file_digest_mapper.as_ref().cloned().unwrap())
     }
 
-    async fn build_snapshotter(&mut self) -> Result<Arc<dyn Snapshotter>> {
-        let snapshotter: Arc<dyn Snapshotter> = match self.configuration.environment {
-            ExecutionEnvironment::Production => {
-                let ongoing_snapshot_directory = self
-                    .configuration
-                    .get_snapshot_dir()?
-                    .join("pending_snapshot");
-
-                let algorithm = match self.configuration.snapshot_compression_algorithm {
-                    CompressionAlgorithm::Gzip => SnapshotterCompressionAlgorithm::Gzip,
-                    CompressionAlgorithm::Zstandard => self
-                        .configuration
-                        .zstandard_parameters
-                        .unwrap_or_default()
-                        .into(),
-                };
-
-                Arc::new(CompressedArchiveSnapshotter::new(
-                    self.configuration.db_directory.clone(),
-                    ongoing_snapshot_directory,
-                    algorithm,
-                    self.root_logger(),
-                )?)
-            }
-            _ => Arc::new(DumbSnapshotter::new()),
-        };
-
-        Ok(snapshotter)
-    }
-
-    /// [Snapshotter] service.
-    pub async fn get_snapshotter(&mut self) -> Result<Arc<dyn Snapshotter>> {
-        if self.snapshotter.is_none() {
-            self.snapshotter = Some(self.build_snapshotter().await?);
-        }
-
-        Ok(self.snapshotter.as_ref().cloned().unwrap())
-    }
-
-    async fn build_certificate_verifier(&mut self) -> Result<Arc<dyn CertificateVerifier>> {
-        let verifier = Arc::new(MithrilCertificateVerifier::new(
-            self.root_logger(),
-            self.get_certificate_repository().await?,
-        ));
-
-        Ok(verifier)
-    }
-
-    /// [CertificateVerifier] service.
-    pub async fn get_certificate_verifier(&mut self) -> Result<Arc<dyn CertificateVerifier>> {
-        if self.certificate_verifier.is_none() {
-            self.certificate_verifier = Some(self.build_certificate_verifier().await?);
-        }
-
-        Ok(self.certificate_verifier.as_ref().cloned().unwrap())
-    }
-
-    async fn build_genesis_verifier(&mut self) -> Result<Arc<ProtocolGenesisVerifier>> {
-        let genesis_verifier: ProtocolGenesisVerifier = match self.configuration.environment {
-            ExecutionEnvironment::Production => ProtocolGenesisVerifier::from_verification_key(
-                ProtocolGenesisVerificationKey::from_json_hex(
-                    &self.configuration.genesis_verification_key,
-                )
-                .map_err(|e| DependenciesBuilderError::Initialization {
-                    message: format!(
-                        "Could not decode hex key to build genesis verifier: '{}'",
-                        self.configuration.genesis_verification_key
-                    ),
-                    error: Some(e),
-                })?,
-            ),
-            _ => ProtocolGenesisSigner::create_deterministic_genesis_signer()
-                .create_genesis_verifier(),
-        };
-
-        Ok(Arc::new(genesis_verifier))
-    }
-
-    /// Return a [ProtocolGenesisVerifier]
-    pub async fn get_genesis_verifier(&mut self) -> Result<Arc<ProtocolGenesisVerifier>> {
-        if self.genesis_verifier.is_none() {
-            self.genesis_verifier = Some(self.build_genesis_verifier().await?);
-        }
-
-        Ok(self.genesis_verifier.as_ref().cloned().unwrap())
-    }
-
-    async fn build_mithril_registerer(&mut self) -> Result<Arc<MithrilSignerRegisterer>> {
-        let registerer = MithrilSignerRegisterer::new(
-            self.get_chain_observer().await?,
-            self.get_verification_key_store().await?,
-            self.get_signer_store().await?,
-            self.configuration.safe_epoch_retention_limit(),
-        );
-
-        Ok(Arc::new(registerer))
-    }
-
-    /// [MithrilSignerRegisterer] service
-    pub async fn get_mithril_registerer(&mut self) -> Result<Arc<MithrilSignerRegisterer>> {
-        if self.mithril_registerer.is_none() {
-            self.mithril_registerer = Some(self.build_mithril_registerer().await?);
-        }
-
-        Ok(self.mithril_registerer.as_ref().cloned().unwrap())
-    }
-
     async fn build_era_reader(&mut self) -> Result<Arc<EraReader>> {
         let era_adapter: Arc<dyn EraReaderAdapter> = match self.configuration.environment {
             ExecutionEnvironment::Production => EraReaderAdapterBuilder::new(
@@ -1155,263 +957,6 @@ impl DependenciesBuilder {
         }
     }
 
-    async fn build_signable_builder_service(&mut self) -> Result<Arc<dyn SignableBuilderService>> {
-        let seed_signable_builder = self.get_signable_seed_builder().await?;
-        let mithril_stake_distribution_builder =
-            Arc::new(MithrilStakeDistributionSignableBuilder::default());
-        let immutable_signable_builder = Arc::new(CardanoImmutableFilesFullSignableBuilder::new(
-            self.get_immutable_digester().await?,
-            &self.configuration.db_directory,
-            self.root_logger(),
-        ));
-        let transactions_importer = self.get_transactions_importer().await?;
-        let block_range_root_retriever = self.get_transaction_repository().await?;
-        let cardano_transactions_builder = Arc::new(CardanoTransactionsSignableBuilder::<
-            MKTreeStoreInMemory,
-        >::new(
-            transactions_importer,
-            block_range_root_retriever,
-        ));
-        let cardano_stake_distribution_builder = Arc::new(
-            CardanoStakeDistributionSignableBuilder::new(self.get_stake_store().await?),
-        );
-        let cardano_database_signable_builder = Arc::new(CardanoDatabaseSignableBuilder::new(
-            self.get_immutable_digester().await?,
-            &self.configuration.db_directory,
-            self.root_logger(),
-        ));
-        let signable_builders_dependencies = SignableBuilderServiceDependencies::new(
-            mithril_stake_distribution_builder,
-            immutable_signable_builder,
-            cardano_transactions_builder,
-            cardano_stake_distribution_builder,
-            cardano_database_signable_builder,
-        );
-        let signable_builder_service = Arc::new(MithrilSignableBuilderService::new(
-            seed_signable_builder,
-            signable_builders_dependencies,
-            self.root_logger(),
-        ));
-
-        Ok(signable_builder_service)
-    }
-
-    /// [SignableBuilderService] service
-    pub async fn get_signable_builder_service(
-        &mut self,
-    ) -> Result<Arc<dyn SignableBuilderService>> {
-        if self.signable_builder_service.is_none() {
-            self.signable_builder_service = Some(self.build_signable_builder_service().await?);
-        }
-
-        Ok(self.signable_builder_service.as_ref().cloned().unwrap())
-    }
-
-    async fn build_signable_seed_builder(&mut self) -> Result<Arc<dyn SignableSeedBuilder>> {
-        let signable_seed_builder_service = Arc::new(AggregatorSignableSeedBuilder::new(
-            self.get_epoch_service().await?,
-        ));
-
-        Ok(signable_seed_builder_service)
-    }
-
-    /// [SignableSeedBuilder] service
-    pub async fn get_signable_seed_builder(&mut self) -> Result<Arc<dyn SignableSeedBuilder>> {
-        if self.signable_seed_builder.is_none() {
-            self.signable_seed_builder = Some(self.build_signable_seed_builder().await?);
-        }
-
-        Ok(self.signable_seed_builder.as_ref().cloned().unwrap())
-    }
-
-    fn build_gcp_uploader(
-        &self,
-        remote_folder_path: CloudRemotePath,
-        allow_overwrite: bool,
-    ) -> Result<GcpUploader> {
-        let logger = self.root_logger();
-        let bucket = self
-            .configuration
-            .snapshot_bucket_name
-            .to_owned()
-            .ok_or_else(|| {
-                DependenciesBuilderError::MissingConfiguration("snapshot_bucket_name".to_string())
-            })?;
-
-        Ok(GcpUploader::new(
-            Arc::new(GcpBackendUploader::try_new(
-                bucket,
-                self.configuration.snapshot_use_cdn_domain,
-                logger.clone(),
-            )?),
-            remote_folder_path,
-            allow_overwrite,
-            FileUploadRetryPolicy::default(),
-        ))
-    }
-
-    fn build_cardano_database_ancillary_uploaders(
-        &self,
-    ) -> Result<Vec<Arc<dyn AncillaryFileUploader>>> {
-        let logger = self.root_logger();
-        if self.configuration.environment == ExecutionEnvironment::Production {
-            match self.configuration.snapshot_uploader_type {
-                SnapshotUploaderType::Gcp => {
-                    let allow_overwrite = true;
-                    let remote_folder_path =
-                        CloudRemotePath::new("cardano-database").join("ancillary");
-
-                    Ok(vec![Arc::new(self.build_gcp_uploader(
-                        remote_folder_path,
-                        allow_overwrite,
-                    )?)])
-                }
-                SnapshotUploaderType::Local => {
-                    let server_url_prefix = self.configuration.get_server_url()?;
-                    let ancillary_url_prefix = server_url_prefix
-                        .sanitize_join(&format!("{CARDANO_DATABASE_DOWNLOAD_PATH}/ancillary/"))?;
-                    let target_dir = self.get_cardano_db_artifacts_dir()?.join("ancillary");
-
-                    std::fs::create_dir_all(&target_dir).map_err(|e| {
-                        DependenciesBuilderError::Initialization {
-                            message: format!("Cannot create '{target_dir:?}' directory."),
-                            error: Some(e.into()),
-                        }
-                    })?;
-
-                    Ok(vec![Arc::new(LocalUploader::new(
-                        ancillary_url_prefix,
-                        &target_dir,
-                        FileUploadRetryPolicy::default(),
-                        logger,
-                    ))])
-                }
-            }
-        } else {
-            Ok(vec![Arc::new(DumbUploader::new(
-                FileUploadRetryPolicy::never(),
-            ))])
-        }
-    }
-
-    fn build_cardano_database_immutable_uploaders(
-        &self,
-    ) -> Result<Vec<Arc<dyn ImmutableFilesUploader>>> {
-        let logger = self.root_logger();
-        if self.configuration.environment == ExecutionEnvironment::Production {
-            match self.configuration.snapshot_uploader_type {
-                SnapshotUploaderType::Gcp => {
-                    let allow_overwrite = false;
-                    let remote_folder_path =
-                        CloudRemotePath::new("cardano-database").join("immutable");
-
-                    Ok(vec![Arc::new(self.build_gcp_uploader(
-                        remote_folder_path,
-                        allow_overwrite,
-                    )?)])
-                }
-                SnapshotUploaderType::Local => {
-                    let server_url_prefix = self.configuration.get_server_url()?;
-                    let immutable_url_prefix = server_url_prefix
-                        .sanitize_join(&format!("{CARDANO_DATABASE_DOWNLOAD_PATH}/immutable/"))?;
-
-                    Ok(vec![Arc::new(LocalUploader::new_without_copy(
-                        immutable_url_prefix,
-                        FileUploadRetryPolicy::default(),
-                        logger,
-                    ))])
-                }
-            }
-        } else {
-            Ok(vec![Arc::new(DumbUploader::new(
-                FileUploadRetryPolicy::never(),
-            ))])
-        }
-    }
-
-    fn build_cardano_database_digests_uploaders(&self) -> Result<Vec<Arc<dyn DigestFileUploader>>> {
-        let logger = self.root_logger();
-        if self.configuration.environment == ExecutionEnvironment::Production {
-            match self.configuration.snapshot_uploader_type {
-                SnapshotUploaderType::Gcp => {
-                    let allow_overwrite = false;
-                    let remote_folder_path =
-                        CloudRemotePath::new("cardano-database").join("digests");
-
-                    Ok(vec![Arc::new(self.build_gcp_uploader(
-                        remote_folder_path,
-                        allow_overwrite,
-                    )?)])
-                }
-                SnapshotUploaderType::Local => {
-                    let server_url_prefix = self.configuration.get_server_url()?;
-                    let digests_url_prefix = server_url_prefix
-                        .sanitize_join(&format!("{CARDANO_DATABASE_DOWNLOAD_PATH}/digests/"))?;
-                    let target_dir = self.get_cardano_db_artifacts_dir()?.join("digests");
-
-                    std::fs::create_dir_all(&target_dir).map_err(|e| {
-                        DependenciesBuilderError::Initialization {
-                            message: format!("Cannot create '{target_dir:?}' directory."),
-                            error: Some(e.into()),
-                        }
-                    })?;
-
-                    Ok(vec![Arc::new(LocalUploader::new(
-                        digests_url_prefix,
-                        &target_dir,
-                        FileUploadRetryPolicy::default(),
-                        logger,
-                    ))])
-                }
-            }
-        } else {
-            Ok(vec![Arc::new(DumbUploader::new(
-                FileUploadRetryPolicy::never(),
-            ))])
-        }
-    }
-
-    async fn build_cardano_database_artifact_builder(
-        &mut self,
-        cardano_node_version: Version,
-    ) -> Result<CardanoDatabaseArtifactBuilder> {
-        let snapshot_dir = self.configuration.get_snapshot_dir()?;
-        let immutable_dir = self.get_cardano_db_artifacts_dir()?.join("immutable");
-
-        let ancillary_builder = Arc::new(AncillaryArtifactBuilder::new(
-            self.build_cardano_database_ancillary_uploaders()?,
-            self.get_snapshotter().await?,
-            self.configuration.get_network()?,
-            self.configuration.snapshot_compression_algorithm,
-            self.root_logger(),
-        )?);
-
-        let immutable_builder = Arc::new(ImmutableArtifactBuilder::new(
-            immutable_dir,
-            self.build_cardano_database_immutable_uploaders()?,
-            self.get_snapshotter().await?,
-            self.configuration.snapshot_compression_algorithm,
-            self.root_logger(),
-        )?);
-
-        let digest_builder = Arc::new(DigestArtifactBuilder::new(
-            self.configuration.get_server_url()?,
-            self.build_cardano_database_digests_uploaders()?,
-            snapshot_dir.join("pending_cardano_database_digests"),
-            self.get_immutable_file_digest_mapper().await?,
-            self.root_logger(),
-        )?);
-
-        Ok(CardanoDatabaseArtifactBuilder::new(
-            self.configuration.db_directory.clone(),
-            &cardano_node_version,
-            self.configuration.snapshot_compression_algorithm,
-            ancillary_builder,
-            immutable_builder,
-            digest_builder,
-        ))
-    }
-
     fn get_cardano_db_artifacts_dir(&self) -> Result<PathBuf> {
         let cardano_db_artifacts_dir = self
             .configuration
@@ -1428,76 +973,6 @@ impl DependenciesBuilder {
         }
 
         Ok(cardano_db_artifacts_dir)
-    }
-
-    async fn build_signed_entity_service(&mut self) -> Result<Arc<dyn SignedEntityService>> {
-        let logger = self.root_logger();
-        let signed_entity_storer = self.build_signed_entity_storer().await?;
-        let epoch_service = self.get_epoch_service().await?;
-        let mithril_stake_distribution_artifact_builder = Arc::new(
-            MithrilStakeDistributionArtifactBuilder::new(epoch_service.clone()),
-        );
-        let snapshotter = self.build_snapshotter().await?;
-        let snapshot_uploader = self.build_snapshot_uploader().await?;
-        let cardano_node_version = Version::parse(&self.configuration.cardano_node_version)
-            .map_err(|e| DependenciesBuilderError::Initialization { message: format!("Could not parse configuration setting 'cardano_node_version' value '{}' as Semver.", self.configuration.cardano_node_version), error: Some(e.into()) })?;
-        let cardano_immutable_files_full_artifact_builder =
-            Arc::new(CardanoImmutableFilesFullArtifactBuilder::new(
-                self.configuration.get_network()?,
-                &cardano_node_version,
-                snapshotter.clone(),
-                snapshot_uploader,
-                self.configuration.snapshot_compression_algorithm,
-                logger.clone(),
-            ));
-        let prover_service = self.get_prover_service().await?;
-        let cardano_transactions_artifact_builder = Arc::new(
-            CardanoTransactionsArtifactBuilder::new(prover_service.clone()),
-        );
-        let stake_store = self.get_stake_store().await?;
-        let cardano_stake_distribution_artifact_builder =
-            Arc::new(CardanoStakeDistributionArtifactBuilder::new(stake_store));
-        let cardano_database_artifact_builder = Arc::new(
-            self.build_cardano_database_artifact_builder(cardano_node_version)
-                .await?,
-        );
-        let dependencies = SignedEntityServiceArtifactsDependencies::new(
-            mithril_stake_distribution_artifact_builder,
-            cardano_immutable_files_full_artifact_builder,
-            cardano_transactions_artifact_builder,
-            cardano_stake_distribution_artifact_builder,
-            cardano_database_artifact_builder,
-        );
-        let signed_entity_service = Arc::new(MithrilSignedEntityService::new(
-            signed_entity_storer,
-            dependencies,
-            self.get_signed_entity_lock().await?,
-            self.get_metrics_service().await?,
-            logger,
-        ));
-
-        // Compute the cache pool for prover service
-        // This is done here to avoid circular dependencies between the prover service and the signed entity service
-        // TODO: Make this part of a warmup phase of the aggregator?
-        if let Some(signed_entity) = signed_entity_service
-            .get_last_cardano_transaction_snapshot()
-            .await?
-        {
-            prover_service
-                .compute_cache(signed_entity.artifact.block_number)
-                .await?;
-        }
-
-        Ok(signed_entity_service)
-    }
-
-    /// [SignedEntityService] service
-    pub async fn get_signed_entity_service(&mut self) -> Result<Arc<dyn SignedEntityService>> {
-        if self.signed_entity_service.is_none() {
-            self.signed_entity_service = Some(self.build_signed_entity_service().await?);
-        }
-
-        Ok(self.signed_entity_service.as_ref().cloned().unwrap())
     }
 
     async fn build_epoch_service(&mut self) -> Result<EpochServiceWrapper> {
@@ -1563,24 +1038,6 @@ impl DependenciesBuilder {
         Ok(self.signed_entity_type_lock.as_ref().cloned().unwrap())
     }
 
-    async fn build_transactions_importer(&mut self) -> Result<Arc<dyn TransactionsImporter>> {
-        let transactions_importer = Arc::new(CardanoTransactionsImporter::new(
-            self.get_block_scanner().await?,
-            self.get_transaction_repository().await?,
-            self.root_logger(),
-        ));
-
-        Ok(transactions_importer)
-    }
-
-    async fn get_transactions_importer(&mut self) -> Result<Arc<dyn TransactionsImporter>> {
-        if self.transactions_importer.is_none() {
-            self.transactions_importer = Some(self.build_transactions_importer().await?);
-        }
-
-        Ok(self.transactions_importer.as_ref().cloned().unwrap())
-    }
-
     async fn build_upkeep_service(&mut self) -> Result<Arc<dyn UpkeepService>> {
         let stake_pool_pruning_task = self.get_stake_store().await?;
         let epoch_settings_pruning_task = self.get_epoch_settings_store().await?;
@@ -1609,27 +1066,6 @@ impl DependenciesBuilder {
         }
 
         Ok(self.upkeep_service.as_ref().cloned().unwrap())
-    }
-
-    async fn build_single_signature_authenticator(
-        &mut self,
-    ) -> Result<Arc<SingleSignatureAuthenticator>> {
-        let authenticator =
-            SingleSignatureAuthenticator::new(self.get_multi_signer().await?, self.root_logger());
-
-        Ok(Arc::new(authenticator))
-    }
-
-    /// [SingleSignatureAuthenticator] service
-    pub async fn get_single_signature_authenticator(
-        &mut self,
-    ) -> Result<Arc<SingleSignatureAuthenticator>> {
-        if self.single_signer_authenticator.is_none() {
-            self.single_signer_authenticator =
-                Some(self.build_single_signature_authenticator().await?);
-        }
-
-        Ok(self.single_signer_authenticator.as_ref().cloned().unwrap())
     }
 
     fn get_epoch_settings_configuration(&mut self) -> Result<AggregatorEpochSettings> {
@@ -1864,50 +1300,6 @@ impl DependenciesBuilder {
         Ok(self.ticker_service.as_ref().cloned().unwrap())
     }
 
-    /// Create [CertifierService] service
-    pub async fn build_certifier_service(&mut self) -> Result<Arc<dyn CertifierService>> {
-        let cardano_network = self.configuration.get_network().with_context(|| {
-            "Dependencies Builder can not get Cardano network while building the chain observer"
-        })?;
-        let sqlite_connection = self.get_sqlite_connection().await?;
-        let open_message_repository = self.get_open_message_repository().await?;
-        let single_signature_repository =
-            Arc::new(SingleSignatureRepository::new(sqlite_connection.clone()));
-        let certificate_repository = self.get_certificate_repository().await?;
-        let certificate_verifier = self.get_certificate_verifier().await?;
-        let genesis_verifier = self.get_genesis_verifier().await?;
-        let multi_signer = self.get_multi_signer().await?;
-        let epoch_service = self.get_epoch_service().await?;
-        let logger = self.root_logger();
-
-        let certifier = Arc::new(MithrilCertifierService::new(
-            cardano_network,
-            open_message_repository,
-            single_signature_repository,
-            certificate_repository,
-            certificate_verifier,
-            genesis_verifier,
-            multi_signer,
-            epoch_service,
-            logger,
-        ));
-
-        Ok(Arc::new(BufferedCertifierService::new(
-            certifier,
-            Arc::new(BufferedSingleSignatureRepository::new(sqlite_connection)),
-            self.root_logger(),
-        )))
-    }
-
-    /// [CertifierService] service
-    pub async fn get_certifier_service(&mut self) -> Result<Arc<dyn CertifierService>> {
-        if self.certifier_service.is_none() {
-            self.certifier_service = Some(self.build_certifier_service().await?);
-        }
-
-        Ok(self.certifier_service.as_ref().cloned().unwrap())
-    }
-
     /// build HTTP message service
     pub async fn build_message_service(&mut self) -> Result<Arc<dyn MessageService>> {
         let certificate_repository = Arc::new(CertificateRepository::new(
@@ -1933,33 +1325,6 @@ impl DependenciesBuilder {
         Ok(self.message_service.as_ref().cloned().unwrap())
     }
 
-    /// Build Prover service
-    pub async fn build_prover_service(&mut self) -> Result<Arc<dyn ProverService>> {
-        let mk_map_pool_size = self
-            .configuration
-            .cardano_transactions_prover_cache_pool_size;
-        let transaction_retriever = self.get_transaction_repository().await?;
-        let block_range_root_retriever = self.get_transaction_repository().await?;
-        let logger = self.root_logger();
-        let prover_service = MithrilProverService::<MKTreeStoreInMemory>::new(
-            transaction_retriever,
-            block_range_root_retriever,
-            mk_map_pool_size,
-            logger,
-        );
-
-        Ok(Arc::new(prover_service))
-    }
-
-    /// [ProverService] service
-    pub async fn get_prover_service(&mut self) -> Result<Arc<dyn ProverService>> {
-        if self.prover_service.is_none() {
-            self.prover_service = Some(self.build_prover_service().await?);
-        }
-
-        Ok(self.prover_service.as_ref().cloned().unwrap())
-    }
-
     /// Remove the dependencies builder from memory to release Arc instances.
     pub async fn vanish(self) {
         self.drop_sqlite_connections().await;
@@ -1975,7 +1340,7 @@ impl DependenciesBuilder {
 
 #[cfg(test)]
 mod tests {
-    use mithril_common::{entities::SignedEntityTypeDiscriminants, test_utils::TempDir};
+    use mithril_common::entities::SignedEntityTypeDiscriminants;
 
     use super::*;
 
@@ -2015,86 +1380,5 @@ mod tests {
             "'is_activated' expected {}, but was {}",
             expected_activation, is_activated
         );
-    }
-
-    mod build_cardano_database_artifact_builder {
-        use super::*;
-
-        #[tokio::test]
-        async fn if_not_local_uploader_create_cardano_database_immutable_dirs() {
-            let snapshot_directory = TempDir::create(
-                "builder",
-                "if_not_local_uploader_create_cardano_database_immutable_dirs",
-            );
-            let cdb_dir = snapshot_directory.join(CARDANO_DB_ARTIFACTS_DIR);
-            let ancillary_dir = cdb_dir.join("ancillary");
-            let immutable_dir = cdb_dir.join("immutable");
-            let digests_dir = cdb_dir.join("digests");
-
-            let mut dep_builder = {
-                let config = Configuration {
-                    snapshot_directory,
-                    // Test environment yield dumb uploaders
-                    environment: ExecutionEnvironment::Test,
-                    ..Configuration::new_sample()
-                };
-
-                DependenciesBuilder::new_with_stdout_logger(config)
-            };
-
-            assert!(!ancillary_dir.exists());
-            assert!(!immutable_dir.exists());
-            assert!(!digests_dir.exists());
-
-            dep_builder
-                .build_cardano_database_artifact_builder(Version::parse("1.0.0").unwrap())
-                .await
-                .unwrap();
-
-            assert!(!ancillary_dir.exists());
-            assert!(immutable_dir.exists());
-            assert!(!digests_dir.exists());
-        }
-
-        #[tokio::test]
-        async fn if_local_uploader_creates_all_cardano_database_subdirs() {
-            let snapshot_directory = TempDir::create(
-                "builder",
-                "if_local_uploader_creates_all_cardano_database_subdirs",
-            );
-            let cdb_dir = snapshot_directory.join(CARDANO_DB_ARTIFACTS_DIR);
-            let ancillary_dir = cdb_dir.join("ancillary");
-            let immutable_dir = cdb_dir.join("immutable");
-            let digests_dir = cdb_dir.join("digests");
-
-            let mut dep_builder = {
-                let config = Configuration {
-                    snapshot_directory,
-                    // Must use production environment to make `snapshot_uploader_type` effective
-                    environment: ExecutionEnvironment::Production,
-                    snapshot_uploader_type: SnapshotUploaderType::Local,
-                    ..Configuration::new_sample()
-                };
-
-                DependenciesBuilder::new_with_stdout_logger(config)
-            };
-            // In production environment the builder can't create in-memory SQLite connections, we
-            // need to provide it manually to avoid creations of unnecessary files.
-            dep_builder.sqlite_connection =
-                Some(Arc::new(ConnectionBuilder::open_memory().build().unwrap()));
-
-            assert!(!ancillary_dir.exists());
-            assert!(!immutable_dir.exists());
-            assert!(!digests_dir.exists());
-
-            dep_builder
-                .build_cardano_database_artifact_builder(Version::parse("1.0.0").unwrap())
-                .await
-                .unwrap();
-
-            assert!(ancillary_dir.exists());
-            assert!(immutable_dir.exists());
-            assert!(digests_dir.exists());
-        }
     }
 }
